Try to sanitize the circle and arc types in SetGeometry.
CID 1267710

npoints is not necessarily 3 points.  There were no checks about
npoints.  More than 3 caused buffer overflows of x and y.  This now
warns for 1 or 2 points and fails for 4 or more points.  If either of
these cases is important someone will have to add code to handle those
cases.  The calculations all appear to assume that there are 3 points.


git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@29886 f0d54148-0727-0410-94bb-9a71ac55c965