Use std::fill to initialize memory of sHeader C arrays in DIPExDataset.
First use of CPL_ARRAYSIZE.  Sadly we cannot yet use the C++11 styles of:

  fill( std::begin(sHeader.ColorTable), std::end(sHeader.ColorTable), 0 );

or initialization with {0} to zero an array.

This is redoing r30413 that was reverted by r30414.  The new code uses
CPL_ARRAYSIZE from r30415 for the first time.  This offers better type
safety and size change safety than the typical memset.  The compiler
then has the option to optimize the code to the equivalent assembly that is
generated by memset (or the vectorized versions of memset).


git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@30416 f0d54148-0727-0410-94bb-9a71ac55c965