Revert back to r31120 for vrtsources.cpp.
The mingw build is failing for isnan.  Better to start over and work
forward to a clean vrtsources.cpp.  r31355 triggered issue #5459
again.  r31357 attempted to fix this, but it isn't obvious what the
proper solution should be.  Likely, the solution should be in cpl_port.h
with CPLIsNan being std::isnan for C++11 and ???



git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@31358 f0d54148-0727-0410-94bb-9a71ac55c965