Implement full read/write support for new NetCDF4 types NC_UBYTE, NC_USHORT, NC_UINT and NC_STRING for variables (except for NC_STRING) and attributes (patch by Julien Demaria, https://github.com/OSGeo/gdal/pull/99)
Only available if NETCDF_HAS_NC4 is defined and only if format=NC4 for writing.
NC_STRING type is supported for reading (scalar and arrays) attributes and is used for writing only for array attributes (scalar are still written as NC_CHAR).
If NETCDF_HAS_NC4 is not defined or format!=NC4, NC_STRING array attributes are written as a single NC_CHAR string using the GDAL {v1,v2,...} convention.



git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@33586 f0d54148-0727-0410-94bb-9a71ac55c965