Fix clang static analyzer complaint in OGCDatumName2EPSGDatumCode().
It appears that the analyzer is not smart enough to see that CSLCount
will return less than 2 if papszTokens[1] would be invalid.  It is
easy enough to check the pointer even if it is a bit redundant.

gt_wkt_srs.cpp:1349:13: warning: Null pointer passed as an argument to a 'nonnull' parameter
        if( EQUAL(papszTokens[1], pszOGCName) )


git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@34123 f0d54148-0727-0410-94bb-9a71ac55c965