Summary: ogrosmlayer.cpp: Don't append garbage data to tag names.
Patch by Jake Molnar.

Bug was introduced in 662729c / r33037.

The string should be terminated at the end of the entire shared
buffer. Rather, it should be terminated at the index after copying stops.

This commit restores the original behaviour before 662729c.

Moving the index variable into the for-loop scope is a good practice in
general, but not in this particular case.

https://github.com/OSGeo/gdal/pull/161/commits/338eca58e193bed68136b0aef7377eb2ff79c96b


git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@35844 f0d54148-0727-0410-94bb-9a71ac55c965