Coordinate transformation: prevent unnecessary coordinate transformations (github #184)
In some cases in which ogr2ogr source and target use the same coordinate systems and
the source and target formats allow for the full input target precision to be preserved
in the output, some coordinates may be altered (e.g. by 1-ulp deviations in floating point numbers)

This is inconvenient for cases where a format is intended for transparent round-trip conversions
(e.g. export data from PostGIS into a Geopackage, then back to PostGIS).

The cause is that longitude/latitude coordinates are always converted internally to radians.
The floating point multiplication by a factor and its inverse may introduce 1-ulp differences.

Here we're adding a test to detect such cases and a fix to avoid conversion to radians when possible.



git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@37138 f0d54148-0727-0410-94bb-9a71ac55c965