Begin IWYU for gmlreader.cpp.
I got a strange crash in travis when I did the entire IWYU before in
r37176.  I was unable to reproduce this crash on any of my systems.
This time, I am going to do the changes progressively.  If there is a
failure, this time I can hopefully figure out what is going on.


git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@37260 f0d54148-0727-0410-94bb-9a71ac55c965