Fix crash in JPGDatasetCommon::ReadICCProfile().
This function trusts the jpeg without validating many of the values.

With a corrupt / malformed input, nICCChunkLength and anChunkSize[i]
can be negative and the file could cause the system to allocate
unreasonably large amounts of ram.  The checks for a nullptr returns
from malloc are only band aids.  ReadICCProfile still has room for
improvement, especially with figuring out what a reasonable maximum
is for nTotalSize.

It was unclear if or how errors should be reported.

Found by autofuzz.


git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@39787 f0d54148-0727-0410-94bb-9a71ac55c965